Move util contents
