+++ src/constants.rs
@@ -110,3 +110,14 @@ pub mod fields {
         FIELD_H8,
     ];
 }
+
+pub mod metadata {
+    pub static METADATA_KEY_RESULT: &str = "Result";
+    pub static METADATA_KEY_EVENT: &str = "Event";
+    pub static METADATA_KEY_SITE: &str = "Site";
+    pub static METADATA_KEY_DATE: &str = "Date";
+    pub static METADATA_KEY_ROUND: &str = "Round";
+    pub static METADATA_KEY_WHITE: &str = "White";
+    pub static METADATA_KEY_BLACK: &str = "Black";
+    pub static METADATA_KEY_FEN: &str = "Fen";
+}
