+++ src/ruleset/classic/util/mod.rs
@@ -1,4 +1,4 @@
+pub(crate) mod castle_data;
 pub(crate) mod move_iterators;
 pub(crate) mod move_legality;
 pub(crate) mod piece_move_iterator;
-pub(crate) mod castle_data;
