--- src/util/metadata.rs
@@ -1,8 +1,0 @@
-pub static METADATA_KEY_RESULT: &str = "Result";
-pub static METADATA_KEY_EVENT: &str = "Event";
-pub static METADATA_KEY_SITE: &str = "Site";
-pub static METADATA_KEY_DATE: &str = "Date";
-pub static METADATA_KEY_ROUND: &str = "Round";
-pub static METADATA_KEY_WHITE: &str = "White";
-pub static METADATA_KEY_BLACK: &str = "Black";
-pub static METADATA_KEY_FEN: &str = "Fen";
