--- src/util/mod.rs
@@ -1,2 +1,0 @@
-pub(crate) mod metadata;
-pub(super) mod parser;
